pyproject.toml
[project]
name = "kazoo"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
  "torch>=2.3",
  "gymnasium>=0.29",
  "pettingzoo>=1.25",
  "numpy>=1.26",
  "pyyaml>=6.0.2",
  "shimmy>=2.0.0",
  "supersuit>=3.10.0",
  "python-dotenv>=1.0.0",
  "requests>=2.31.0",
  "tqdm>=4.65.0",
  "omegaconf>=2.3.0",
  "hydra-core>=1.3.0",
  "stable-baselines3>=2.3.0",
  "torch-geometric>=2.5.0",
  "scikit-learn>=1.7.0",
]

[tool.uv]             
package = true         

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.11.6",
]
